refactor: move paths.ts to utils/, clean up redundant comments
Move src/paths.ts and its test to src/utils/ alongside other utility
modules. Strip redundant JSDoc, section dividers, and issue references.
