+++ src/agent/pi/session-factory.ts
@@ -18,7 +18,7 @@ import type { SandboxExecutor } from "../middleware/executor.js";
 import { createAgentTools } from "../tools/index.js";
 import { overrideSessionSystemPrompt } from "./system-prompt-override.js";
 import { buildAgentSystemPrompt } from "../workspace/context.js";
-import { resolveAgentWorkspacePath } from "../../paths.js";
+import { resolveAgentWorkspacePath } from "../../utils/paths.js";
 
 const ISOTOPES_HOME = process.env.ISOTOPES_HOME || path.join(process.env.HOME || "/tmp", ".isotopes");
 const DEFAULT_MODEL = "claude-opus-4-7";
