+++ src/utils/paths.test.ts
